There's something I'm missing with implementing context manager in C. Pushed it to python land instead.
Simple script to segfault python via double-free bug.

	import dvdread

	d = dvdread.DVD('/dev/sr0')
	with d:
		d.Open()

	print(d.Path)

Upon conclusion of the with code block, DVD_dealloc is called, which should not happen. But, after d.Path is evaluated then dealloc is called again, which is what formally leads to the segfault. I have no idea why the with block is leading to a dealloc. This problem exists if doing with dvdread.DVD('/dev/sr0') as d style as well.